🔨 Refactor: 텍스트 뷰어 공통 컴포넌트화
