+++ src/types/RichTextViewer/RichTextViewer.type.ts
@@ -1,0 +1,3 @@
+export interface RichTextViewerProps {
+  content: string;
+}
